solana_transaction_error

Enum TransactionError

pub enum TransactionError {

    AccountInUse,
    AccountLoadedTwice,
    AccountNotFound,
    ProgramAccountNotFound,
    InsufficientFundsForFee,
    InvalidAccountForFee,
    AlreadyProcessed,
    BlockhashNotFound,
    InstructionError(u8, InstructionError),
    CallChainTooDeep,
    MissingSignatureForFee,
    InvalidAccountIndex,
    SignatureFailure,
    InvalidProgramForExecution,
    SanitizeFailure,
    ClusterMaintenance,
    AccountBorrowOutstanding,
    WouldExceedMaxBlockCostLimit,
    UnsupportedVersion,
    InvalidWritableAccount,
    WouldExceedMaxAccountCostLimit,
    WouldExceedAccountDataBlockLimit,
    TooManyAccountLocks,
    AddressLookupTableNotFound,
    InvalidAddressLookupTableOwner,
    InvalidAddressLookupTableData,
    InvalidAddressLookupTableIndex,
    InvalidRentPayingAccount,
    WouldExceedMaxVoteCostLimit,
    WouldExceedAccountDataTotalLimit,
    DuplicateInstruction(u8),
    InsufficientFundsForRent {
        account_index: u8,
    },
    MaxLoadedAccountsDataSizeExceeded,
    InvalidLoadedAccountsDataSizeLimit,
    ResanitizationNeeded,
    ProgramExecutionTemporarilyRestricted {
        account_index: u8,
    },
    UnbalancedTransaction,
    ProgramCacheHitMaxLimit,
    CommitCancelled,

}

Reasons a transaction might be rejected.

Variants

AccountInUse

An account is already being processed in another transaction in a way that does not support parallelism

AccountLoadedTwice

A Pubkey appears twice in the transactionโs account_keys. Instructions can reference Pubkeys more than once but the message must contain a list with no duplicate keys

AccountNotFound

Attempt to debit an account but found no record of a prior credit.

ProgramAccountNotFound

Attempt to load a program that does not exist

InsufficientFundsForFee

The from Pubkey does not have sufficient balance to pay the fee to schedule the transaction

InvalidAccountForFee

This account may not be used to pay transaction fees

AlreadyProcessed

The bank has seen this transaction before. This can occur under normal operation when a UDP packet is duplicated, as a user error from a client not updating its recent_blockhash, or as a double-spend attack.

BlockhashNotFound

The bank has not seen the given recent_blockhash or the transaction is too old and the recent_blockhash has been discarded.

InstructionError(u8, InstructionError)

An error occurred while processing an instruction. The first element of the tuple indicates the instruction index in which the error occurred.

CallChainTooDeep

Loader call chain is too deep

MissingSignatureForFee

Transaction requires a fee but has no signature present

InvalidAccountIndex

Transaction contains an invalid account reference

SignatureFailure

Transaction did not pass signature verification

InvalidProgramForExecution

This program may not be used for executing instructions

SanitizeFailure

Transaction failed to sanitize accounts offsets correctly implies that account locks are not taken for this TX, and should not be unlocked.

ClusterMaintenance

AccountBorrowOutstanding

Transaction processing left an account with an outstanding borrowed reference

WouldExceedMaxBlockCostLimit

Transaction would exceed max Block Cost Limit

UnsupportedVersion

Transaction version is unsupported

InvalidWritableAccount

Transaction loads a writable account that cannot be written

WouldExceedMaxAccountCostLimit

Transaction would exceed max account limit within the block

WouldExceedAccountDataBlockLimit

Transaction would exceed account data limit within the block

TooManyAccountLocks

Transaction locked too many accounts

AddressLookupTableNotFound

Address lookup table not found

InvalidAddressLookupTableOwner

Attempted to lookup addresses from an account owned by the wrong program

InvalidAddressLookupTableData

Attempted to lookup addresses from an invalid account

InvalidAddressLookupTableIndex

Address table lookup uses an invalid index

InvalidRentPayingAccount

Transaction leaves an account with a lower balance than rent-exempt minimum

WouldExceedMaxVoteCostLimit

Transaction would exceed max Vote Cost Limit

WouldExceedAccountDataTotalLimit

Transaction would exceed total account data limit

DuplicateInstruction(u8)

Transaction contains a duplicate instruction that is not allowed

InsufficientFundsForRent

Transaction results in an account with insufficient funds for rent

Fields

account_index: u8

MaxLoadedAccountsDataSizeExceeded

Transaction exceeded max loaded accounts data size cap

InvalidLoadedAccountsDataSizeLimit

LoadedAccountsDataSizeLimit set for transaction must be greater than 0.

ResanitizationNeeded

Sanitized transaction differed before/after feature activiation. Needs to be resanitized.

ProgramExecutionTemporarilyRestricted

Program execution is temporarily restricted on an account.

Fields

account_index: u8

UnbalancedTransaction

The total balance before the transaction does not equal the total balance after the transaction

ProgramCacheHitMaxLimit

Program cache hit max limit.

CommitCancelled

Commit cancelled internally.